+++ src/js/core/Core.js
@@ -380,6 +380,8 @@ export const onCall = async (message: CoreMessage): Promise<void> => {
             // wait for self-release and then carry on
             await device.waitForFirstRun();
         } else {
+            // cancel popup request
+            postMessage(new UiMessage(POPUP.CANCEL_POPUP_REQUEST));
             postMessage(new ResponseMessage(responseID, false, { error: ERROR.DEVICE_CALL_IN_PROGRESS.message }));
             throw ERROR.DEVICE_CALL_IN_PROGRESS;
         }
@@ -395,7 +397,9 @@ export const onCall = async (message: CoreMessage): Promise<void> => {
     // device is available
     // set public variables, listeners and run method
     /* eslint-disable no-use-before-define */
-    device.on(DEVICE.BUTTON, onDeviceButtonHandler);
+    device.on(DEVICE.BUTTON, (device, code) => {
+        onDeviceButtonHandler(device, code, method);
+    });
     device.on(DEVICE.PIN, onDevicePinHandler);
     device.on(DEVICE.PASSPHRASE, method.useEmptyPassphrase ? onEmptyPassphraseHandler : onDevicePassphraseHandler);
     device.on(DEVICE.PASSPHRASE_ON_DEVICE, () => {
@@ -616,12 +620,19 @@ const closePopup = (): void => {
  * @returns {Promise<void>}
  * @memberof Core
  */
-const onDeviceButtonHandler = async (device: Device, code: string): Promise<void> => {
+const onDeviceButtonHandler = async (device: Device, code: string, method: AbstractMethod): Promise<void> => {
     // wait for popup handshake
-    await getPopupPromise().promise;
+    const addressRequest = code === 'ButtonRequest_Address';
+    if (!addressRequest || (addressRequest && method.useUi)) {
+        await getPopupPromise().promise;
+    }
+    const data = typeof method.getButtonRequestData === 'function' ? method.getButtonRequestData(code) : null;
     // request view
     postMessage(new DeviceMessage(DEVICE.BUTTON, { device: device.toMessageObject(), code: code }));
-    postMessage(new UiMessage(UI.REQUEST_BUTTON, { device: device.toMessageObject(), code: code }));
+    postMessage(new UiMessage(UI.REQUEST_BUTTON, { device: device.toMessageObject(), code: code, data }));
+    if (addressRequest && !method.useUi) {
+        postMessage(new UiMessage(UI.ADDRESS_VALIDATION, data));
+    }
 };
 
 /**