+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# feature/address-validation
+__added__
+- Added `UI.ADDRESS_VALIDATION` event
+
+__changed__
+- `getAddress`, `cardanoGetAddress`, `ethereumGetAddress`, `liskGetAddress`, `nemGetAddress`, `rippleGetAddress`, `stellarGetAddress`, `tezosGetAddress` methods (allow to handle `UI.ADDRESS_VALIDATION` event)
+- refactor `ButtonRequest_Address` view in popup: display address and copy to clipboard button
+
 # 6.0.3
 __added__
 - Added `TrezorConnect.tezosGetAddress` method